requirements.txt
websocket-client~=1.2.3